TravelManagement/src/main/java/com/crossover/project/entities/JqDataTable.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package com.crossover.project.entities;

import java.util.List;

/**
 *
 * @author waqas.ali2
 */
public class JqDataTable {

    private int recordsTotal;
    private int recordsFiltered;
    private int sEcho;
    private int iDisplayLength;
    private int draw;
    private List aaData;

    public int getRecordsTotal() {
        return recordsTotal;
    }

    public void setRecordsTotal(int recordsTotal) {
        this.recordsTotal = recordsTotal;
    }

    public int getRecordsFiltered() {
        return recordsFiltered;
    }

    public void setRecordsFiltered(int recordsFiltered) {
        this.recordsFiltered = recordsFiltered;
    }

    public int getsEcho() {
        return sEcho;
    }

    public void setsEcho(int sEcho) {
        this.sEcho = sEcho;
    }

    public int getiDisplayLength() {
        return iDisplayLength;
    }

    public void setiDisplayLength(int iDisplayLength) {
        this.iDisplayLength = iDisplayLength;
    }

    public int getDraw() {
        return draw;
    }

    public void setDraw(int draw) {
        this.draw = draw;
    }

    public List getAaData() {
        return aaData;
    }

    public void setAaData(List aaData) {
        this.aaData = aaData;
    }
}
